requirements.txt
streamlit 
azure-cognitiveservices-vision-customvision